match template naming convention
