+++ packages/spruce-templates/src/interfaces/Autoloader.ts
@@ -1,0 +1,13 @@
+export interface IAutoloadedImport {
+	[path: string]: {
+		defaultImport?: string
+		namedImports?: string[]
+	}
+}
+
+export interface IAutoloader {
+	nameCamel: string
+	namePascal: string
+	namePascalSingular: string
+	imports: IAutoloadedImport
+}
